test: update imports and refactor test cases for AboutUs, CooperationSection, NewsSection
