+++ app/[lang]/page.test.tsx
@@ -8,8 +8,12 @@ jest.mock('next/navigation', () => ({
 }));
 
 describe('AboutUs Main Page', () => {
-  it('should redirect to /about-us on mount', () => {
-    render(<AboutUs />);
+  it('should redirect to /about-us on mount', async () => {
+    const mockParams = Promise.resolve({ lang: 'uk' as const });
+
+    const ResolvedPage = await AboutUs({ params: mockParams });
+    render(ResolvedPage);
+
     expect(redirect).toHaveBeenCalledWith('/about-us');
   });
 });
